RAIN-6906 : DSS Added landing page breadcrumb to the module (#2865)
* RAIN-6906 : DSS Added landing page breadcrumb to the module

* RAIN-6906 : Removed the unwanted changes

* RAIN-6906 : Corrected the redirectiong URL for state and national

* RAIN-7024 OBPS :: Added Architect using login details and code for FI options (#2796)

* RAIN-7024 Added Architect using login details

* RAIN-7024 Added code for FI options

* RAIN-6906 : Given the condition for redirction url for side bar icon … (#2878)

* RAIN-6906 : Given the condition for redirction url for side bar icon for state and national dashboard

* Update EmployeeSideBar.js


* RAIN-7019 OBPS eDCR Scrutiny :: Applications are not listed when tried to search from different page in employee inbox (#2884)

* Frontend Mono UI Build issue resolved by fixing the peer dependency version of datatables package (#2885) (#2889)

* Update package.json

* updated build by clearing

* removed data tables dependency

* Revert "removed data tables dependency"

This reverts commit b8b8661afefe35d087238f6407de43695bb4e4f1.

* Corrected package dependency issues for data tables

* removed unnesceesary change

* Revert "Update Dockerfile"

This reverts commit 9825dad78fbf9c012ace2c5dfc0e2cc58fefc2bd.

* removed other changes

* updated alignment

* RAIN-7003 OBPS All flows documents issues fixed, multiple API calls fixed, alignment issues, functional issue and new components for reusing the document codes (#2892)

* FIX on doc issue

* UI : OBPS Doc issue fixed

* RAIN-4760 : Documents Re upload same file issue fixed

* DOCS details hooks created to show

* inprogreess

* updated documents stufff

* updated documents stufff

* RAIN-7003 All flows documents issues fixed, multiple api calls fixed and new components for reusing the document codes

* RAIN-7003 removed log


* Removed the MAP API Key

+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/obps/src/pages/citizen/BPASendToArchitect/index.js
@@ -11,7 +11,7 @@ const getPath = (path, params) => {
   return path;
 }
 
-const getBPAEditDetails = (data, APIScrutinyDetails,mdmsData,nocdata,t) => {
+const getBPAEditDetails = async (data, APIScrutinyDetails,mdmsData,nocdata,t) => {
 
   const getBlockIds = (unit) => {
     let blocks = {};
@@ -101,7 +101,7 @@ const getBPAEditDetails = (data, APIScrutinyDetails,mdmsData,nocdata,t) => {
     applicationType:data?.additionalDetails?.applicationType || APIScrutinyDetails?.appliactionType,
     serviceType:data?.additionalDetails?.serviceType || APIScrutinyDetails?.applicationSubType
   }
-
+  sessionStorage.setItem("BPA_IS_ALREADY_WENT_OFF_DETAILS", JSON.stringify(true));
   return data;
 }
 
@@ -141,17 +141,21 @@ const BPASendToArchitect = ({ parentRoute }) => {
 
   const editApplication = window.location.href.includes("editApplication");
 
-  useEffect(() => {
-     application = bpaData ? bpaData[0]:{};
-     if (data1 && nocdata) {
-      application = bpaData[0];
-       if (editApplication) {
-         application.isEditApplication = true;
-       }
-       sessionStorage.setItem("bpaInitialObject", JSON.stringify({ ...application }));
-       let bpaEditDetails = getBPAEditDetails(application,data1,mdmsData,nocdata,t);
-       setParams({ ...params, ...bpaEditDetails });
-     }
+  useEffect(async () => {
+    let isAlready = sessionStorage.getItem("BPA_IS_ALREADY_WENT_OFF_DETAILS");
+    isAlready = isAlready ? JSON.parse(isAlready) : true;
+    if (!isAlready && !isNocLoading && !isBpaSearchLoading && !isLoading) {
+      application = bpaData ? bpaData[0]:{};
+      if (data1 && nocdata) {
+       application = bpaData[0];
+        if (editApplication) {
+          application.isEditApplication = true;
+        }
+        sessionStorage.setItem("bpaInitialObject", JSON.stringify({ ...application }));
+        let bpaEditDetails = await getBPAEditDetails(application,data1,mdmsData,nocdata,t);
+        setParams({ ...params, ...bpaEditDetails });
+      }
+    }
   }, [bpaData,data1,mdmsData,nocdata]);
 
 
@@ -195,6 +199,10 @@ const BPASendToArchitect = ({ parentRoute }) => {
 
   const CheckPage = Digit?.ComponentRegistryService?.getComponent('BPACheckPage') ;
   const OBPSAcknowledgement = Digit?.ComponentRegistryService?.getComponent('BPAAcknowledgement');
+
+  if (isNocLoading || isBpaSearchLoading || isLoading) {
+    return <Loader />
+  }
 
   return (
     <Switch>
